chore: refactored image-queue. updated name getAuthenticatedClientAuthenticationUtil.
