+++ backend/src/services/lens/create-onchain-comment-typed-data.lens.service.ts
@@ -1,4 +1,4 @@
-import { getAuthenticatedClientUtil } from "../../utils/authentication/get-authenticated-client.authentication.util";
+import { getAuthenticatedClientAuthenticationUtil } from "../../utils/authentication/get-authenticated-client.authentication.util";
 import createOnchainCommentTypedDataMutationGraphql from "../../graphql/mutations/create-onchain-comment-typed-data.mutation.graphql";
 import type { OnchainCommentRequest } from "../../gql/graphql";
 
@@ -6,7 +6,7 @@ const createOnchainCommentTypedDataService = async (
   request: OnchainCommentRequest
 ) => {
   console.log("createOnchainCommentTypedDataLensService request", request);
-  const client = await getAuthenticatedClientUtil();
+  const client = await getAuthenticatedClientAuthenticationUtil();
   const result = await client
     .mutation(createOnchainCommentTypedDataMutationGraphql, { request })
     .toPromise();
